requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.2.2=pypi_0
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.10.11=pypi_0
aiosignal=1.3.1=pypi_0
async-timeout=5.0.1=pypi_0
attrs=25.3.0=pypi_0
ca-certificates=2025.1.31=hbd8a1cb_1
certifi=2025.1.31=pypi_0
charset-normalizer=3.4.1=pypi_0
click=7.1.2=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
datasets=3.1.0=pypi_0
dill=0.3.8=pypi_0
evaluate=0.4.3=pypi_0
filelock=3.16.1=pypi_0
fonttools=4.57.0=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.9.0=pypi_0
fuzzy-c-means=1.6.3=pypi_0
huggingface-hub=0.30.2=pypi_0
idna=3.10=pypi_0
importlib-resources=6.4.5=pypi_0
joblib=1.4.2=pypi_0
jsonlines=4.0.0=pypi_0
kiwisolver=1.4.7=pypi_0
ld_impl_linux-64=2.43=h712a8e2_4
libffi=3.2.1=he1b5a44_1007
libgcc=14.2.0=h767d61c_2
libgcc-ng=14.2.0=h69a702a_2
libgomp=14.2.0=h767d61c_2
liblzma=5.8.1=hb9d3cd8_0
liblzma-devel=5.8.1=hb9d3cd8_0
libretranslatepy=2.1.1=pypi_0
libsqlite=3.46.0=hde9e2c9_0
libstdcxx=14.2.0=h8f9b012_2
libstdcxx-ng=14.2.0=h4852527_2
libzlib=1.2.13=h4ab18f5_6
loguru=0.7.3=pypi_0
lxml=5.4.0=pypi_0
matplotlib=3.7.5=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.5=h2d0b736_3
nltk=3.9.1=pypi_0
numpy=1.24.4=pypi_0
openssl=1.1.1w=hd590300_0
packaging=25.0=pypi_0
pandas=2.0.3=pypi_0
pillow=10.4.0=pypi_0
pip=24.3.1=pyh8b19718_0
portalocker=3.0.0=pypi_0
propcache=0.2.0=pypi_0
pyarrow=17.0.0=pypi_0
pydantic=1.10.21=pypi_0
pyparsing=3.1.4=pypi_0
python=3.8.0=h357f687_5
python-dateutil=2.9.0.post0=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h8c095d6_2
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rouge-score=0.1.2=pypi_0
sacrebleu=2.5.1=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
setuptools=75.3.0=pyhd8ed1ab_0
six=1.17.0=pypi_0
sqlite=3.46.0=h6d4b2fc_0
tabulate=0.8.10=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.20.3=pypi_0
torch=1.8.0=pypi_0
torchtext=0.9.0=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.46.3=pypi_0
translate-toolkit=3.13.5=pypi_0
typer=0.3.2=pypi_0
typing-extensions=4.13.2=pypi_0
tzdata=2025.2=pypi_0
urllib3=2.2.3=pypi_0
wcwidth=0.2.13=pypi_0
wheel=0.45.1=pyhd8ed1ab_0
xxhash=3.5.0=pypi_0
xz=5.8.1=hbcc6ac9_0
xz-gpl-tools=5.8.1=hbcc6ac9_0
xz-tools=5.8.1=hb9d3cd8_0
yarl=1.15.2=pypi_0
zipp=3.20.2=pypi_0
zlib=1.2.13=h4ab18f5_6
